# Node definition manager refactor