--- a/src/content_mapblock.h
+++ b/src/content_mapblock.h
@@ -49,7 +49,7 @@ public:
 	MeshMakeData *data;
 	MeshCollector *collector;
 
-	INodeDefManager *nodedef;
+	const NodeDefManager *nodedef;
 	scene::IMeshManipulator *meshmanip;
 
 // options