… src/mapnode.cpp
… src/mapnode.h
… src/nodedef.cpp
… src/nodedef.h
… src/objdef.h
… src/pathfinder.cpp
… src/reflowscan.cpp
… src/reflowscan.h
… src/rollback_interface.cpp
… src/script/common/c_content.cpp
… src/script/common/c_content.h
… src/script/cpp_api/s_client.cpp
… src/script/cpp_api/s_node.cpp
… src/script/cpp_api/s_nodemeta.cpp
… src/script/lua_api/l_client.cpp
… src/script/lua_api/l_env.cpp
… src/script/lua_api/l_item.cpp
… src/script/lua_api/l_mapgen.cpp
… src/script/lua_api/l_vmanip.cpp
… src/server.cpp
… src/server.h
… src/serverenvironment.cpp
… src/unittest/test.cpp
… src/unittest/test_mapnode.cpp
… src/unittest/test_noderesolver.cpp
… src/unittest/test_schematic.cpp
… src/unittest/test_voxelalgorithms.cpp
… src/unittest/test_voxelmanipulator.cpp
… src/voxel.cpp
… src/voxel.h