This is mainly a maintenance PR, put there is a smaller performance improvement too.

There was three classes associated with the node definition manager:

- `INodeDefManager` in `nodedef.h`: a wrapper interface for reading the manager
- `IWritableNodeDefManager` in `nodedef.h`: a wrapper interface for modifying the manager
- `CNodeDefManager`.cpp in `nodedef.cpp`: actual implementation

I moved these three classes into one `NodeDefManager` class. This eliminates the virtual function calls generated by the interfaces, slightly increasing performance where `INodeDefManager::get()` was frequently called. But more importantly, making the manager easier to understand. Now new contributors will find the implementation more easily. The changes on an image:
![nodedef](https://user-images.githubusercontent.com/20600448/35781718-f2d45bb0-09ed-11e8-9e1f-e4dd0e1056a3.png)

Performance difference:

  | mapgen v7 | mesh generation | liquid reflow scan | /fixlight here 32
-- | -- | -- | -- | --
Current | 137238,5 μs/chunk | 2143,5 μs/block | 71,1 μs/block | 125321,1 μs
This PR | 137348,4 μs/chunk | 2105,3 μs/block | 49,0 μs/block | 110190,9 μs
Change | 0,08% | -1,78% | -31,08% | -12,07%

I made the following steps to achieve this, in each step you should be able to compile the source code:

1. Rename `IWritableNodeDefManager` to `NodeDefManager`: this interface contains all methods that are used elsewhere, this will be the base of the new class.
2. Make `INodeDefManager` functions constant: there are listeners that listen on node registration completion. Even a constant pointer to `NodeDefManager` should be able to register listeners (the actual content of the manager remains constant). Therefore, I made the listener array `mutable`.
3. Replace all `INodeDefManager` pointers with `const NodeDefManager` pointers
4. Remove `INodeDefManager`: it is not used anymore.
5. Merge `NodeDefManager` with `CNodeDefManager`: move data fields and inline methods to the header file. `NodeDefManager` has no more virtual members.
6. Add doxygen documentation: make the manager contributor-friendly.

The possible downsides that I see:

- Conflicts with other PRs: fortunately resolving the conflicts is easy: just replace `*INodeDefManager` with `const *NodeDefManager`.
- Breaks consistency with `IItemDefManager` and `ICraftDefManager`: maybe those should be refactored too.
- If the implementation is changed (a new member will be added to the class), lots of files will need recompiling: the structure of the node definition manager is changed rarely.