@SmallJoker 2018-02-07
I like the idea of combining these classes with a single one that offers the same functions and restrictions. Concept approval from my side.